+++ database/migrations/2025_05_09_000004_create_products_table.php
@@ -4,22 +4,21 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up(): void
     {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->foreignId('category_id')
-                  ->nullable()
-                  ->constrained('categories')
-                  ->nullOnDelete()
-                  ->cascadeOnUpdate();
+                ->nullable()
+                ->constrained('categories')
+                ->nullOnDelete()
+                ->cascadeOnUpdate();
             $table->string('name', 150);
             $table->string('sku', 50)->unique()->nullable();
             $table->text('description')->nullable();
-            $table->decimal('price', 15, 2);
-            $table->decimal('cost_price', 15, 2)->nullable();
+            $table->unsignedBigInteger('price');
+            $table->unsignedBigInteger('cost_price')->nullable();
             $table->string('image_url', 255)->nullable();
             $table->boolean('has_variants')->default(false);
             $table->boolean('is_active')->default(true);
