+++ database/migrations/2025_05_09_000010_create_discounts_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up(): void
     {
         Schema::create('discounts', function (Blueprint $table) {
@@ -13,8 +12,8 @@ return new class extends Migration
             $table->string('code', 50)->unique();
             $table->string('name', 100)->nullable();
             $table->string('type', 15)->comment('percentage | fixed');
-            $table->decimal('value', 15, 2);
-            $table->decimal('min_purchase', 15, 2)->default(0);
+            $table->unsignedBigInteger('value');
+            $table->unsignedBigInteger('min_purchase')->default(0);
             $table->integer('max_uses')->nullable()->comment('null = unlimited');
             $table->integer('used_count')->default(0);
             $table->timestamp('valid_from')->nullable();
